fix: Change pdf_width_pts and pdf_height_pts to float type
- pypdfium2 returns float values from page.get_size()
- Pydantic validation was failing on fractional parts
- Changed TaskPageResponse fields from Optional[int] to Optional[float]
- Changed database columns from Integer to Float
- Updated migration script to use FLOAT type

+++ database_migration.py
@@ -22,8 +22,8 @@ def migrate_database(db_path: str) -> None:
 
         # Columns to add
         columns_to_add = [
-            ("pdf_width_pts", "INTEGER"),
-            ("pdf_height_pts", "INTEGER"),
+            ("pdf_width_pts", "FLOAT"),
+            ("pdf_height_pts", "FLOAT"),
             ("render_scale", "FLOAT"),
         ]
